Some Changes

+++ src/com/metrostate/ics460/project2/sender/DataSender.java
@@ -4,6 +4,6 @@ package com.metrostate.ics460.project2.sender;
 
 public interface DataSender {
 
-	public void sendData(byte[] bytes, int packetSize, long timeout, String ipAddress, int port, int windowSize);
+	public void sendData(byte[] bytes, int packetSize, int timeout, String ipAddress, int port, int windowSize);
 	
 }
